not final version

+++ src/main/java/com/example/demo/Repository/SkillsRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.Repository;
 import com.example.demo.Entity.Skills;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkillsRepository extends JpaRepository<Skills, Integer> {
+public interface SkillsRepository extends JpaRepository<Skills, Long> {
 }
